requirements.txt
annotated-types==0.7.0
anyio==4.11.0
appdirs==1.4.4
blinker==1.8.2
cachetools==6.2.0
certifi==2025.8.3
# Editable Git install with no remote (cfclient==2024.11.post1.dev11+gb2cf5f2)
-e c:\windows\system32\crazyflie-clients-python
cflib==0.1.27
charset-normalizer==3.4.3
click==8.1.7
colorama==0.4.6
earthengine-api==1.6.10
fastapi==0.118.0
Flask==3.0.3
freetype-py==2.5.1
google-ai-generativelanguage==0.6.15
google-api-core==2.25.2
google-api-python-client==2.184.0
google-auth==2.41.1
google-auth-httplib2==0.2.0
google-cloud-core==2.4.3
google-cloud-storage==3.4.0
google-crc32c==1.7.1
google-generativeai==0.8.5
google-resumable-media==2.7.2
googleapis-common-protos==1.70.0
grpcio==1.75.1
grpcio-status==1.71.2
h11==0.16.0
hsluv==5.0.4
httplib2==0.31.0
idna==3.10
iniconfig==2.0.0
itsdangerous==2.2.0
Jinja2==3.1.4
kiwisolver==1.4.7
libusb-package==1.0.26.1
MarkupSafe==2.1.5
numpy==1.26.4
packaging==24.2
pandas==2.3.3
pluggy==1.5.0
proto-plus==1.26.1
protobuf==5.29.5
pyasn1==0.6.1
pyasn1_modules==0.4.2
pydantic==2.11.10
pydantic_core==2.33.2
PyOpenGL==3.1.7
pyparsing==3.2.5
PyQt6==6.7.1
PyQt6-Qt6==6.7.3
PyQt6_sip==13.9.0
pyqtgraph==0.13.7
PySDL2==0.9.16
pysdl2-dll==2.24.0
pyserial==3.5
pytest==8.3.4
python-dateutil==2.9.0.post0
pytz==2025.2
pyusb==1.2.1
PyYAML==6.0.2
pyzmq==26.2.0
requests==2.32.5
rsa==4.9.1
scipy==1.14.1
setuptools==75.6.0
six==1.17.0
sniffio==1.3.1
starlette==0.48.0
tqdm==4.67.1
typing-inspection==0.4.2
typing_extensions==4.15.0
tzdata==2025.2
uritemplate==4.2.0
urllib3==2.5.0
uvicorn==0.37.0
vispy==0.14.3
Werkzeug==3.0.3
